// ecis-nodejs-zgwx-feat_240603/service/account-zgxw/ecissdk/model/data.ts
export interface MysqlSelResp {
  result: string
  data: { rows: any[] }
  msg?: string
}